refactor main file
